Move SANGraphHead to components/grit_layers.py
layers.py now only re-exports; all layer implementations live in
dedicated modules.
